refactor(gh-pages): rename the file with the Search Box demo component
