+++ pages/src/navigation/navigation.ts
@@ -1,1 +1,2 @@
 export * from './sidebar/sidebar';
+export * from './header/header-search';
